Move PV deletion to diskmaker controllers and remove deleter
Serialize deletion and creation of PV's like local-static-provisioner
does to avoid competing delete / create requests. The LV controller and
LVS controller will handle their respective PV deletion, and the deleter
controller can be completely removed.
